Move f32 to wave_common to fix circular imports
