+++ corrscope/wave_common.py
@@ -1,0 +1,3 @@
+import numpy as np
+
+f32 = np.single
